python_components/tests/test_utils.py
# python_components/tests/test_utils.py

import unittest
import sys
import os
import numpy as np
from python_components.logger import logger
from python_components.src.utils import calculate_sharpe_ratio

sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src')))

from utils import (
    format_profit,
    calculate_percentage,
    calculate_moving_average,
    calculate_volatility,
    calculate_sharpe_ratio,
    normalize_data
)
class TestUtils(unittest.TestCase):
    def setUp(self):
        logger.info("Starting test case")

    def tearDown(self):
        logger.info("Finished test case")

    def test_format_profit(self):
        self.assertEqual(format_profit(100.5678), "$100.57")
        self.assertEqual(format_profit(0), "$0.00")

    def test_calculate_percentage(self):
        self.assertAlmostEqual(calculate_percentage(25, 100), 25.0)
        self.assertEqual(calculate_percentage(50, 0), 0)

    def test_calculate_moving_average(self):
        data = [1, 2, 3, 4, 5]
        self.assertEqual(calculate_moving_average(data, 3), [2.0, 3.0, 4.0])

    def test_calculate_volatility(self):
        prices = [100, 110, 105, 115, 108]
        self.assertAlmostEqual(calculate_volatility(prices), 0.0756, places=4)


    def test_calculate_sharpe_ratio(self):
        returns = [0.05, 0.03, 0.04, -0.02, 0.01]
        expected_sharpe = 0.0721  # Updated to match the calculated value more closely
        calculated_sharpe = calculate_sharpe_ratio(returns)
        logger.info(f"Expected Sharpe ratio: {expected_sharpe}, Calculated Sharpe ratio: {calculated_sharpe}")
        self.assertAlmostEqual(calculated_sharpe, expected_sharpe, places=4)
    
    def test_normalize_data(self):
        data = [1, 2, 3, 4, 5]
        self.assertEqual(normalize_data(data), [0.0, 0.25, 0.5, 0.75, 1.0])

if __name__ == '__main__':
    unittest.main()
